requirements.txt
asgiref==3.3.1
autopep8==1.5.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.0
Django==3.1.7
django-braces==1.14.0
django-filter==2.4.0
django-oauth-toolkit==1.4.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.2
idna==2.10
Markdown==3.3.4
oauthlib==3.1.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
